+++ action_server/docs/CHANGELOG.md
@@ -2,6 +2,13 @@
 
 ## Unreleased
 
+- Add top level command: `oauth2`, providing OAuth2 related utilities
+- Add `sema4ai-config` subcommand for `oauth2`
+  - Returns OAuth2 configuration for Sema4.ai provided OAuth2 applications 
+- Add `user-config-path` subcommand for `oauth2`
+  - Returns the path to user's local OAuth config file
+  - `--json` argument can be provided to get the result in JSON format
+
 ## 0.18.0 - 2024-08-27
 
 - Add to the package metadata an `action_package_version` field and rename `version` to `metadata_version`.